+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to napari-js are documented here. The format roughly follows
 [Keep a Changelog](https://keepachangelog.com/); versions follow [SemVer](https://semver.org/).
 
-## [Unreleased]
+## [0.12.0]
 
 ### Added